Added logging outputs, and submoduled oh-my-zsh.
logs directory has gitkeep file to ensure it stays, but all output files are ignored.

Logging is split into relevant logs, so it is easier to parse.

Removed the shell command to install oh-my-zsh because it is now a submodule. The repo
is symbolically linked to the home directory, where zshrc soource's it.